% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingTemplate.R
\docType{package}
\name{testingTemplate}
\alias{testingTemplate}
\alias{testingTemplate-package}
\title{This Package Helps Illustrate Template Tricks}
\description{
Here we have collected a set of tricks or pieces
of code which we have found useful in formatting
output reports in cleaner, more standardized ways.
}
